@@ -5,7 +5,7 @@ using JsonApiDotNetCore.Internal.Contracts;
 using JsonApiDotNetCore.Middleware;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Internal;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 
@@ -14,29 +14,34 @@ namespace JsonApiDotNetCore.Extensions
     // ReSharper disable once InconsistentNaming
     public static class IApplicationBuilderExtensions
     {
+        /// <summary>
+        /// Adds necessary components such as routing to your application
+        /// </summary>
+        /// <param name="app"></param>
+        /// <param name="useMvc"></param>
+        /// <returns></returns>
         public static IApplicationBuilder UseJsonApi(this IApplicationBuilder app, bool useMvc = true)
         {
             DisableDetailedErrorsIfProduction(app);
             LogResourceGraphValidations(app);
-            app.UseEndpointRouting();
             app.UseMiddleware<CurrentRequestMiddleware>();
+            app.UseRouting();
             if (useMvc)
+            {
                 app.UseMvc();
-
+            }
             using (var scope = app.ApplicationServices.CreateScope())
             {
                 var inverseRelationshipResolver = scope.ServiceProvider.GetService<IInverseRelationships>();
                 inverseRelationshipResolver?.Resolve();
             }
-
             return app;
         }
 
         private static void DisableDetailedErrorsIfProduction(IApplicationBuilder app)
         {
-            var environment = (IHostingEnvironment)app.ApplicationServices.GetService(typeof(IHostingEnvironment));
-
-            if (environment.IsProduction())
+            var webHostEnvironment = (IWebHostEnvironment) app.ApplicationServices.GetService(typeof(IWebHostEnvironment));
+            if (webHostEnvironment.EnvironmentName == "Production")
             {
                 JsonApiOptions.DisableErrorStackTraces = true;
                 JsonApiOptions.DisableErrorSource = true;
